+++ lib/pipedrive/resources/lead_label.rb
@@ -3,5 +3,6 @@
 module Pipedrive
   class LeadLabel < Resource
     self.resources_url = "leadLabels"
+    update_method :patch
   end
 end